프로젝트 정보 저장 기능 개선 및 redis 서버 연결
