+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/dto/CreateProjectResponseDTO.java
@@ -5,12 +5,12 @@ package com.sk.skala.axcalibur.feature.dto;
 /**
  * 프로젝트 생성 응답 DTO (IF-PR-0004)
  */
-public class CreateProjectResponse {
+public class CreateProjectResponseDTO {
 
     private String projectId;
     private String avalon;
 
-    public CreateProjectResponse(String projectId, String avalon) {
+    public CreateProjectResponseDTO(String projectId, String avalon) {
         this.projectId = projectId;
         this.avalon = avalon;
     }
