+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/dto/DeleteProjectResponseDTO.java
@@ -2,9 +2,9 @@ package com.sk.skala.axcalibur.feature.dto;
 
 // 프로젝트 삭제 응답 DTO (IF-PR-0003, IF-PR-0005)
 // 응답 본문이 비어있음 (HTTP Status Code로 성공 여부 판단)
-public class DeleteProjectResponse {
+public class DeleteProjectResponseDTO {
     
     // 기본 생성자
-    public DeleteProjectResponse() {}
+    public DeleteProjectResponseDTO() {}
     
 }
