+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/dto/RequirementInfoDTO.java
@@ -1,6 +1,6 @@
 package com.sk.skala.axcalibur.feature.dto;
 
-public class RequirementInfo {
+public class RequirementInfoDTO {
     private Long requirementId; // 요구사항 PK
     private String name;        // 요구사항 이름
     private String desc;        // 요구사항 설명
@@ -9,9 +9,9 @@ public class RequirementInfo {
     private String middle;      // 요구사항 중분류
     private String minor;       // 요구사항 소분류
 
-    public RequirementInfo() {}
+    public RequirementInfoDTO() {}
 
-    public RequirementInfo(Long requirementId, String name, String desc, String priority, String major, String middle, String minor) {
+    public RequirementInfoDTO(Long requirementId, String name, String desc, String priority, String major, String middle, String minor) {
         this.requirementId = requirementId;
         this.name = name;
         this.desc = desc;
