+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/dto/SaveProjectResponseDTO.java
@@ -4,15 +4,15 @@ import java.time.LocalDateTime;
 
 // 프로젝트 저장 응답 DTO (IF-PR-0001)
 // 설계서 기준: requestTime 응답
-public class SaveProjectResponse {
+public class SaveProjectResponseDTO {
     
     private LocalDateTime requestTime;   // 요청 생성 시간
     
     // 기본 생성자
-    public SaveProjectResponse() {}
+    public SaveProjectResponseDTO() {}
     
     // 생성자
-    public SaveProjectResponse(LocalDateTime requestTime) {
+    public SaveProjectResponseDTO(LocalDateTime requestTime) {
         this.requestTime = requestTime;
     }
