+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/entity/CategoryEntity.java
@@ -22,7 +22,7 @@ import lombok.Setter;
 @Builder
 @Entity
 @Table(name = "category")
-public class Category {
+public class CategoryEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
