+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/entity/PriorityEntity.java
@@ -22,18 +22,18 @@ import lombok.Setter;
 @AllArgsConstructor
 @Builder
 @Entity
-@Table(name = "request_middle", uniqueConstraints = {
+@Table(name = "priority", uniqueConstraints = {
     @UniqueConstraint(columnNames = {"name"})
 })
-public class RequestMiddle {
+public class PriorityEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "`key`")   
-    private Integer key;        // 중분류 키 (PK, AUTO_INCREMENT)
+    @Column(name = "`key`")
+    private Integer key;        // 중요도 키 (PK, AUTO_INCREMENT)
 
-    @Column(name = "name", unique = true, nullable = false, length = 20)
-    private String name;                 // 중분류 명 (UNIQUE)
+    @Column(name = "name", unique = true, nullable = false, length = 50)
+    private String name;                 // 중요도 명 (UNIQUE)
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;     // 생성 일자
