+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/entity/ProjectEntity.java
@@ -21,7 +21,7 @@ import lombok.Setter;
 @Builder
 @Entity
 @Table(name = "project")
-public class Project {
+public class ProjectEntity {
 
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
