+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/entity/RequestMinorEntity.java
@@ -25,7 +25,7 @@ import lombok.Setter;
 @Table(name = "request_minor", uniqueConstraints = {
     @UniqueConstraint(columnNames = {"name"})
 })
-public class RequestMinor {
+public class RequestMinorEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
