+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/repository/CategoryRepository.java
@@ -5,13 +5,13 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sk.skala.axcalibur.feature.entity.Category;
+import com.sk.skala.axcalibur.feature.entity.CategoryEntity;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
     
     // 카테고리명으로 조회
-    Optional<Category> findByName(String name);
+    Optional<CategoryEntity> findByName(String name);
 
     // 카테고리명 존재 여부 확인
     boolean existsByName(String name);
