+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/repository/ContextRepository.java
@@ -5,13 +5,13 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sk.skala.axcalibur.feature.entity.Context;
+import com.sk.skala.axcalibur.feature.entity.ContextEntity;
 
 @Repository
-public interface ContextRepository extends JpaRepository<Context, Integer> {
+public interface ContextRepository extends JpaRepository<ContextEntity, Integer> {
     
     // 컨텍스트명으로 조회
-    Optional<Context> findByName(String name);
+    Optional<ContextEntity> findByName(String name);
 
     // 컨텍스트명 존재 여부 확인
     boolean existsByName(String name);
