+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/repository/FilePathRepository.java
@@ -5,19 +5,19 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sk.skala.axcalibur.feature.entity.FilePath;
-import com.sk.skala.axcalibur.feature.entity.Project;
+import com.sk.skala.axcalibur.feature.entity.FilePathEntity;
+import com.sk.skala.axcalibur.feature.entity.ProjectEntity;
 
 @Repository
-public interface FilePathRepository extends JpaRepository<FilePath, Integer> {
+public interface FilePathRepository extends JpaRepository<FilePathEntity, Integer> {
     
     // 프로젝트별 파일 경로 목록 조회
-    List<FilePath> findByProjectKey(Project projectKey);
+    List<FilePathEntity> findByProjectKey(ProjectEntity projectKey);
 
     // 특정 경로로 조회
-    List<FilePath> findByPath(String path);
+    List<FilePathEntity> findByPath(String path);
 
     // 경로에 특정 문자열 포함된 것 조회
-    List<FilePath> findByPathContaining(String path);
+    List<FilePathEntity> findByPathContaining(String path);
 
 }
