+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/repository/ProjectRepository.java
@@ -5,17 +5,14 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sk.skala.axcalibur.feature.entity.Project;
+import com.sk.skala.axcalibur.feature.entity.ProjectEntity;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<ProjectEntity, Integer> {
 
     // 프로젝트 ID로 조회 (비즈니스 키)
-    Optional<Project> findById(String id);
+    Optional<ProjectEntity> findById(String id);
 
     // 프로젝트 ID 존재 여부
     boolean existsById(String id);
-
-    // avalon 값으로 프로젝트 조회
-    Optional<Project> findByAvalon(String avalon);
 }
