+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/repository/RequestRepository.java
@@ -6,22 +6,22 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sk.skala.axcalibur.feature.entity.Priority;
-import com.sk.skala.axcalibur.feature.entity.Project;
-import com.sk.skala.axcalibur.feature.entity.Request;
+import com.sk.skala.axcalibur.feature.entity.PriorityEntity;
+import com.sk.skala.axcalibur.feature.entity.ProjectEntity;
+import com.sk.skala.axcalibur.feature.entity.RequestEntity;
 
 @Repository
-public interface RequestRepository extends JpaRepository<Request, Integer> {
-
-    // 요구사항 명으로 조회 (Unique)
-    Optional<Request> findByName(String name);
+public interface RequestRepository extends JpaRepository<RequestEntity, Integer> {
 
     // 프로젝트별 요구사항 목록 조회
-    List<Request> findByProjectKey(Project projectKey);
+    List<RequestEntity> findByProjectKey(ProjectEntity projectKey);
 
     // 우선순위별 요구사항 목록 조회
-    List<Request> findByPriorityKey(Priority priorityKey);
+    List<RequestEntity> findByPriorityKey(PriorityEntity priorityKey);
 
-    // 요구사항명 존재 여부 확인
-    boolean existsByName(String name);
+    // 프로젝트별 중복 체크
+    boolean existsByProjectKeyAndName(ProjectEntity projectKey, String name);
+
+    // 프로젝트별 요구사항 조회
+    Optional<RequestEntity> findByProjectKeyAndName(ProjectEntity projectKey, String name);
 }
